installed sqlALCHEMY added sqlalchemy to the requirements.txt restructured files in to package in TV_bot created routes for the views created models to host the db logic initialized necessary imports in the init file.
