+++ py/TV_bot/requirements.txt
@@ -3,30 +3,53 @@ aiohttp==3.6.2
 asciichart==0.1
 async-timeout==3.0.1
 attrs==19.3.0
+autobahn==21.2.1
+Automat==20.2.0
 ccxt==1.33.17
 certifi==2020.6.20
 cffi==1.14.4
 chardet==3.0.4
 click==7.1.2
+colorama==0.4.4
 config==0.5.0.post0
+constantly==15.1.0
 cryptography==3.3.1
+dateparser==1.0.0
 Flask==1.1.2
 flask-ngrok==0.0.25
+Flask-SQLAlchemy==2.4.4
+hyperlink==21.0.0
 idna==2.10
+incremental==17.5.0
 itsdangerous==1.1.0
 Jinja2==2.11.2
 MarkupSafe==1.1.1
 multidict==4.7.6
 numpy==1.19.1
 pandas==1.2.1
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
 pycares==3.1.1
 pycparser==2.20
+PyHamcrest==2.0.2
+pyOpenSSL==20.0.1
+python-binance==0.7.9
 python-dateutil==2.8.1
 pytz==2020.1
 redis==3.5.3
+regex==2020.11.13
 requests==2.24.0
 rq==1.7.0
+service-identity==18.1.0
 six==1.15.0
+SQLAlchemy==1.3.23
+termcolor==1.1.0
+Twisted==20.3.0
+txaio==20.12.1
+tzlocal==2.1
+ujson==4.0.2
 urllib3==1.25.10
+websockets==8.1
 Werkzeug==1.0.1
 yarl==1.1.0
+zope.interface==5.2.0
