+++ py/TV_bot/templates/trade_history.html
