+++ py/run.py
@@ -1,0 +1,6 @@
+from TV_bot import app
+
+if __name__ == '__main__':
+    "instantiate the flask app in debug mode"
+    app.debug = True
+    app.run()
